# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DARK TIERS Minecraft PvP Ranking Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user or null
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
...
  /announcements:
    get:
      operationId: listAnnouncements
      tags: [announcements]
      summary: List announcements
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Announcement list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnnouncementList"
    post:
      operationId: createAnnouncement
      tags: [announcements]
      summary: Create announcement (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementInput"
      responses:
        "201":
          description: Created announcement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
  /announcements/{id}:
    get:
      operationId: getAnnouncement
      tags: [announcements]
      summary: Get announcement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Announcement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
    patch:
      operationId: updateAnnouncement
      tags: [announcements]
      summary: Update announcement (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementUpdate"
      responses:
        "200":
          description: Updated announcement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
    delete:
      operationId: deleteAnnouncement
      tags: [announcements]
      summary: Delete announcement (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ADMIN
...
  /admin/users:
    get:
      operationId: listUsers
      tags: [admin]
      summary: List all users (admin)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserList"
  /admin/users/{id}:
    patch:
      operationId: updateUser
      tags: [admin]
      summary: Update user (role, suspend) (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/analytics:
    get:
      operationId: getAnalytics
      tags: [admin]
      summary: Analytics data for charts
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [7d, 30d, 90d]
      responses:
        "200":
          description: Analytics data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsData"
  /admin/players/{id}/reset-rating:
    post:
      operationId: resetPlayerRating
      tags: [admin]
      summary: Reset player rating (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetRatingInput"
      responses:
        "200":
          description: Reset complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
  /admin/players/{id}/change-tier:
    post:
      operationId: changePlayerTier
      tags: [admin]
      summary: Change player tier (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeTierInput"
      responses:
        "200":
          description: Tier changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
components:
  schemas:
...
    AuthUser:
      type: object
      required: [id, discordId, username, role]
      properties:
        id:
          type: integer
        discordId:
          type: string
        username:
          type: string
        displayName:
          type: ["string", "null"]
        avatar:
          type: ["string", "null"]
        role:
          type: string
          enum: [owner, admin, moderator, tester, user]
        isSuspended:
          type: boolean
        createdAt:
          type: string
...
    Player:
      type: object
      required: [id, username, uuid, region, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        uuid:
          type: string
        region:
          type: string
        country:
          type: ["string", "null"]
        userId:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
...
    Announcement:
      type: object
      required: [id, title, content, type, authorId, authorName, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        type:
          type: string
          enum: [news, update, event, maintenance]
        isPinned:
          type: boolean
        authorId:
          type: integer
        authorName:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
    AnnouncementInput:
      type: object
      required: [title, content, type, authorId]
      properties:
        title:
          type: string
        content:
          type: string
        type:
          type: string
          enum: [news, update, event, maintenance]
        isPinned:
          type: boolean
        authorId:
          type: integer
    AnnouncementUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        type:
          type: string
          enum: [news, update, event, maintenance]
        isPinned:
          type: boolean
    AnnouncementList:
      type: object
      required: [announcements, total, page, limit]
      properties:
        announcements:
          type: array
          items:
            $ref: "#/components/schemas/Announcement"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    User:
      type: object
      required: [id, discordId, username, role, isSuspended, createdAt]
      properties:
        id:
          type: integer
        discordId:
          type: string
        username:
          type: string
        displayName:
          type: ["string", "null"]
        avatar:
          type: ["string", "null"]
        role:
          type: string
          enum: [owner, admin, moderator, tester, user]
        isSuspended:
          type: boolean
        playerId:
          type: ["integer", "null"]
        createdAt:
          type: string
    UserUpdate:
      type: object
      properties:
        role:
          type: string
          enum: [owner, admin, moderator, tester, user]
        isSuspended:
          type: boolean
        playerId:
          type: integer
    UserList:
      type: object
      required: [users, total, page, limit]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    AnalyticsData:
      type: object
      required: [userGrowth, matchActivity, tierDistribution, gamemodePopularity]
      properties:
        userGrowth:
          type: array
          items:
            $ref: "#/components/schemas/TimeSeriesPoint"
        matchActivity:
          type: array
          items:
            $ref: "#/components/schemas/TimeSeriesPoint"
        tierDistribution:
          type: array
          items:
            $ref: "#/components/schemas/CategoryPoint"
        gamemodePopularity:
          type: array
          items:
            $ref: "#/components/schemas/CategoryPoint"
    TimeSeriesPoint:
      type: object
      required: [date, value]
      properties:
        date:
          type: string
        value:
          type: integer
    CategoryPoint:
      type: object
      required: [label, value]
      properties:
        label:
          type: string
        value:
          type: integer
    ResetRatingInput:
      type: object
      required: [gamemodeId]
      properties:
        gamemodeId:
          type: integer
        newRating:
          type: integer
    ChangeTierInput:
      type: object
      required: [gamemodeId, tierId]
      properties:
        gamemodeId:
          type: integer
        tierId:
          type: integer