# lib/api-zod/src/generated/types/announcementUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DARK TIERS Minecraft PvP Ranking Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AnnouncementUpdateType } from './announcementUpdateType';

export interface AnnouncementUpdate {
  title?: string;
  content?: string;
  type?: AnnouncementUpdateType;
  isPinned?: boolean;
}